refactor(web): split websocket health view helpers
