+++ web/src/lib/features/machines/components/machine-health-panel-view.ts
@@ -1,13 +1,8 @@
 import { i18nStore } from '$lib/i18n/store.svelte'
 import { friendlyTransportLabel } from '../machine-setup'
 import { formatMachineRelativeTime } from '../machine-i18n'
-import type {
-  MachineCLIStatus,
-  MachineSnapshot,
-  MachineWebsocketHealth,
-  MachineWebsocketHealthLayer,
-  WebsocketHealthState,
-} from '../types'
+import type { MachineCLIStatus, MachineSnapshot, WebsocketHealthState } from '../types'
+import { buildWebsocketLevelCards } from './machine-websocket-health-view'
 
 export type HealthStatCard = {
   label: string
@@ -197,50 +192,6 @@ export function buildLevelCards(snapshot: MachineSnapshot): HealthLevelCard[] {
   ]
 }
 
-function buildWebsocketLevelCards(health: MachineWebsocketHealth): HealthLevelCard[] {
-  return [
-    websocketLevelCard(
-      'l2',
-      i18nStore.t('machines.machineHealthPanel.levels.l2Link'),
-      health.l2,
-      i18nStore.t('machines.machineHealthPanel.dynamic.noWebsocketLinkObservation'),
-    ),
-    websocketLevelCard(
-      'l3',
-      i18nStore.t('machines.machineHealthPanel.levels.l3ControlPlaneNetwork'),
-      health.l3,
-      i18nStore.t('machines.machineHealthPanel.dynamic.noWebsocketNetworkObservation'),
-    ),
-    websocketLevelCard(
-      'l4',
-      i18nStore.t('machines.machineHealthPanel.levels.l4WebsocketTransport'),
-      health.l4,
-      i18nStore.t('machines.machineHealthPanel.dynamic.noWebsocketTransportObservation'),
-    ),
-    websocketLevelCard(
-      'l5',
-      i18nStore.t('machines.machineHealthPanel.levels.l5MachineAgent'),
-      health.l5,
-      i18nStore.t('machines.machineHealthPanel.dynamic.noWebsocketApplicationObservation'),
-    ),
-  ]
-}
-
-function websocketLevelCard(
-  id: string,
-  label: string,
-  layer: MachineWebsocketHealthLayer | undefined,
-  fallbackValue: string,
-): HealthLevelCard {
-  return {
-    id,
-    label,
-    state: mapWebsocketLayerState(layer?.state),
-    value: layer?.reason ?? fallbackValue,
-    meta: checkedAtLabel(layer?.observedAt),
-  }
-}
-
 export function buildAuditRows(snapshot: MachineSnapshot): HealthAuditRow[] {
   if (!snapshot.fullAudit) {
     return []
@@ -346,18 +297,5 @@ export function stateLabel(state: string): string {
       return i18nStore.t('machines.machineHealthPanel.status.failed')
     default:
       return i18nStore.t('machines.machineHealthPanel.status.unknown')
-  }
-}
-
-function mapWebsocketLayerState(state: WebsocketHealthState | undefined): string {
-  switch (state) {
-    case 'healthy':
-      return 'ok'
-    case 'degraded':
-      return 'warn'
-    case 'failed':
-      return 'error'
-    default:
-      return 'unknown'
   }
 }
